(BSR) feat(etl_jobs): make export parallel (#4144)
