+++ jobs/etl_jobs/external/encrypted_exports/requirements.in
@@ -4,3 +4,4 @@ boto3
 botocore
 typer
 google-cloud-secret-manager==2.20.0
+loguru
