+++ jobs/etl_jobs/external/encrypted_exports/requirements.txt
@@ -61,6 +61,8 @@ jmespath==1.0.1
     # via
     #   boto3
     #   botocore
+loguru==0.7.3
+    # via -r requirements.in
 markdown-it-py==3.0.0
     # via rich
 mdurl==0.1.2
@@ -77,6 +79,8 @@ protobuf==4.25.6
     #   grpc-google-iam-v1
     #   grpcio-status
     #   proto-plus
+psutil==7.0.0
+    # via -r requirements.in
 pyasn1==0.6.1
     # via
     #   pyasn1-modules
@@ -104,9 +108,7 @@ six==1.17.0
 typer==0.15.1
     # via -r requirements.in
 typing-extensions==4.12.2
-    # via
-    #   rich
-    #   typer
+    # via typer
 urllib3==2.3.0
     # via
     #   botocore
